nova_creative_demo/requirements.txt
streamlit==1.36.0
boto3==1.37.2
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.3
pillow==10.2.0
moviepy==2.1.2  # For video processing
opencv-python-headless==4.11.0.86
tabulate==0.9.0
termcolor==2.5.0
pyyaml
pydantic==2.10.2
